test(flip): restore browser interaction and measurement coverage
